refaktorált adatbázis, cors készítés

+++ src/main/java/com/gymprojekt/forevergym/repository/RoleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface RoleRepository extends CrudRepository<Role, Long> {
+public interface RoleRepository extends CrudRepository<Role, Integer> {
 
     Optional<Role> findByRoleName(String roleName);
